requirements.txt
pyelftools==0.28
construct==2.10.*
